整合一下

+++ src/main/java/com/ols/ols_project/controller/TaskController.java
@@ -60,7 +60,6 @@ public class TaskController {
     @Autowired
     private UidGenService uidGenService;
 
-
     @Value("${image.fileFath}")
     private String desFilePath;
 
@@ -493,18 +492,20 @@ public class TaskController {
         String [] fileNameArray= fileNameArray0.split(",");
         String delFailure="";
         String delThumbFailure="";
-        for(int i=0;i<fileNameArray.length;i++){
-            File fromFile = new File(path+fileNameArray[i]);//找到文件
-            File fromThumbFile = new File(thumbPath+fileNameArray[i]);
-            if (fromFile.exists()) {
-                fromFile.delete();
-            }else{
-                delFailure+=fileNameArray[i]+",";
-            }
-            if (fromThumbFile.exists()) {
-                fromThumbFile.delete();//遍历删除文件夹及其子内容
-            }else{
-                delThumbFailure+=fileNameArray[i]+",";
+        if(fileNameArray.length!=0){
+            for(int i=0;i<fileNameArray.length;i++){
+                File fromFile = new File(path+fileNameArray[i]);//找到文件
+                File fromThumbFile = new File(thumbPath+fileNameArray[i]);
+                if (fromFile.exists()) {
+                    fromFile.delete();
+                }else{
+                    delFailure+=fileNameArray[i]+",";
+                }
+                if (fromThumbFile.exists()) {
+                    fromThumbFile.delete();//遍历删除文件夹及其子内容
+                }else{
+                    delThumbFailure+=fileNameArray[i]+",";
+                }
             }
         }
         if(delFailure!=""){
@@ -558,9 +559,9 @@ public class TaskController {
                              @RequestParam(value = "field") String field,
                              @RequestParam(value = "order") String order){
         String result= JSON.toJSONStringWithDateFormat(
-                new Result(
-                        taskService.getAllTask(query, pageNum, pageSize,queryInfo,searchType,searchInfo,field,order)
-                        ,"0"
+                        new Result(
+                                taskService.getAllTask(query, pageNum, pageSize,queryInfo,searchType,searchInfo,field,order)
+                                ,"0"
                         ,"获取所有任务成功")
                 ,"yyyy-MM-dd hh:mm:ss"
                 ,SerializerFeature.WriteNonStringValueAsString);
@@ -652,6 +653,60 @@ public class TaskController {
         zipFile.delete();//临时压缩包用后删除
     }
 
+    @GetMapping("/downloadDocTask")
+    public void downloadDocTask(@RequestParam(value = "taskId") long taskId,
+                           HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        //判断环境
+        String str="";
+        String os = System.getProperty("os.name");
+        if (os.toLowerCase().startsWith("win")) {
+            str = "\\";
+        }else if(os.toLowerCase().startsWith("lin")){
+            str = "/";
+        }
+        //获取要下载的所有文件名
+        JSONArray fileNameArray = taskService.getDocFileNameByTaskId(taskId);
+        String path = desFilePath;//规定的下载地址
+        //String path = "http:\\\\yuyy.info\\image\\ols\\";//linux 路径有点问题
+        //新建临时文件夹用于打包所有要下载的问文件 若已存在，则先删除再创建
+        File fileAllTemp = new File(path+taskId);
+        if (fileAllTemp.exists()) {
+            FileUtils.deleteFile(fileAllTemp);//遍历删除文件夹及其子内容
+            fileAllTemp.mkdirs();
+        }else {
+            fileAllTemp.mkdirs();
+        }
+        for(int i=0;i<fileNameArray.size();i++){
+            File fromFile = new File(path+fileNameArray.get(i));//找到文件
+            File toFile = new File(path+taskId+str+fileNameArray.get(i));//目标文件地址 用于将要打包的文件放在一起
+            copy(fromFile.toPath(),toFile.toPath());//将源文件复制进临时文件夹  用于打包
+        }
+
+        FileOutputStream zip = new FileOutputStream(new File(fileAllTemp.getPath()+".zip"));
+        ZipUtils.toZip(fileAllTemp.getPath(),zip,true);
+
+        File zipFile = new File(fileAllTemp.getPath()+".zip");//选中压缩文件
+        if(zipFile.exists())
+        {
+            response.setContentType("application/x-msdownload"); //设置响应类型,此处为下载类型
+            response.setHeader("Content-Disposition", "attachment;filename=\""+taskId+".zip\"");//以附件的形式打开
+            InputStream inputStream = new FileInputStream(zipFile);
+            ServletOutputStream outputStream = response.getOutputStream();
+            byte b[] = new byte[1024];
+            int n;
+            while((n = inputStream.read(b)) != -1)
+            {
+                outputStream.write(b,0,n);
+            }
+            outputStream.close();
+            inputStream.close();
+        }else{
+            request.setAttribute("result", "文件不存在！下载失败！");
+        }
+        FileUtils.deleteFile(fileAllTemp);//临时文件夹用完删除
+        zipFile.delete();//临时压缩包用后删除
+    }
+
     @GetMapping("/getAdminImgChartData")
     public String getAdminImgChartData(
             @RequestParam("year") String year
@@ -692,7 +747,6 @@ public class TaskController {
                 "yyyy-MM-dd");
         return result;
     }
-
     @GetMapping("/getAllReleaseDocById")
     public String getAllReleaseDocById(
             @RequestParam("year") String year,
@@ -706,7 +760,6 @@ public class TaskController {
                 "yyyy-MM-dd");
         return result;
     }
-
 
     @GetMapping(value = "/getAcceptImgTaskByUserId")
     public String getAcceptImgTaskByUserId(
@@ -722,6 +775,28 @@ public class TaskController {
         String result = JSON.toJSONStringWithDateFormat(
                 new Result(
                         taskService.getAcceptImgTaskByUserId(
+                                Long.parseLong(userId), query, pageNum, pageSize, queryInfo, searchInfo,field,order)
+                        , "0", "获取已接受任务成功")
+                , "yyyy-MM-dd hh:mm:ss"
+                , SerializerFeature.WriteNonStringValueAsString
+        );
+        return result;
+    }
+
+    @GetMapping(value = "/getAcceptDocTaskByUserId")
+    public String getAcceptDocTaskByUserId(
+            @RequestParam(value = "userId") String userId,
+            @RequestParam(value = "query") String query,
+            @RequestParam(value = "page") Integer pageNum,
+            @RequestParam(value = "limit") Integer pageSize,
+            @RequestParam(value = "queryInfo") String queryInfo,
+            @RequestParam(value = "searchInfo") String searchInfo,
+            @RequestParam(value = "field") String field,
+            @RequestParam(value = "order") String order
+    ) {
+        String result = JSON.toJSONStringWithDateFormat(
+                new Result(
+                        taskService.getAcceptDocTaskByUserId(
                                 Long.parseLong(userId), query, pageNum, pageSize, queryInfo, searchInfo,field,order)
                         , "0", "获取已接受任务成功")
                 , "yyyy-MM-dd hh:mm:ss"
@@ -750,6 +825,47 @@ public class TaskController {
                 , "yyyy-MM-dd hh:mm:ss"
                 , SerializerFeature.WriteNonStringValueAsString
         );
+        return result;
+    }
+
+    @GetMapping(value = "/getReleaseDocTaskByUserId")
+    public String getReleaseDocTaskByUserId(
+            @RequestParam(value = "userId") String userId,
+            @RequestParam(value = "query") String query,
+            @RequestParam(value = "page") Integer pageNum,
+            @RequestParam(value = "limit") Integer pageSize,
+            @RequestParam(value = "queryInfo") String queryInfo,
+            @RequestParam(value = "searchInfo") String searchInfo,
+            @RequestParam(value = "field") String field,
+            @RequestParam(value = "order") String order
+    ) {
+        // layui默认数据表格的status为0才显示数据
+        String result = JSON.toJSONStringWithDateFormat(
+                new Result(
+                        taskService.getReleaseDocTaskByUserId(Long.parseLong(userId), query, pageNum, pageSize, queryInfo, searchInfo,field,order)
+                        , "0"
+                        , "获取已发布任务成功")
+                , "yyyy-MM-dd hh:mm:ss"
+                , SerializerFeature.WriteNonStringValueAsString
+        );
+        return result;
+    }
+    @GetMapping(value = "getDiskInfo")
+    public String getDiskInfo(){
+        HashMap<String,Object> data=new HashMap<>();
+        File file = new File("D:\\images\\");
+
+        int gByte = 1024 * 1024 * 1024;
+        long total = file.getTotalSpace() / gByte;
+        long used = (file.getTotalSpace()- file.getFreeSpace())/ gByte;
+        long free = file.getFreeSpace() / gByte;
+        data.put("total",total);
+        data.put("free",free);//
+        data.put("used",used);//已用
+        String result= JSON.toJSONStringWithDateFormat(
+                new Result(data,"200","获取个人已发布成功"),
+                "yyyy-MM-dd");
+        System.out.println(result);
         return result;
     }
 }
